requirements.txt

fastapi
uvicorn[standard]
requests
pydantic-settings
google-generativeai
google-auth
jinja2